Change: Add max-width and edit width of image
By adding a max width property to the image element in the header, I was
able to restrict the growth of the element which was throwing the page
out of whack a little when zooming. I then edited the width/flex-basis
of the element to make the basis of the image a bit bigger for improved
flexbox functionality.