Feat: Copy header html from index to contact.html
After making the header for the homepage, I've decided to use the same
page/container structure and layout for my About and Contact pages. the
idea is that the header template from index.html will form the body
content as it is sufficient for the amount of page content that my
design calls for.

There are/will be several differences from the original code, but I
decided to commit the original code copy-pasted to highlight changes in
my next commit.

One that I will highlight here is that I will have to return to add the
footer in once I design it, and at that stage I will center/justify the
page content vertically as opposed to the way that content is vertically
justified in the index.html file. This applies to the About page too as
the template will be the same, except without the call to action button.