Change: Edit html template to accommodate page
As mentioned in my last commit, these are the changes that I've made to
the contact page html. I've edited the text in the <h1> element, and
I've edited the text in the <p> element in the same parent container.

I've edited the text embedded in the link element within the button to
reflect its new function; I switched the Spotify external link to
mailto:alexanderbrightmanmusic@gmail.com so that people can open an
email form with their default client with my email address pre-filled.

I also swapped the image link out for a different image in ~/images.

It's worth noting that I haven't changed the names of any class or id
selectors, so the css in styles.css should translate very well to
contact.html, and subsequently about.html when I get to that. I
anticipate that one or two flexbox properties will need some tinkering.