Change: Edit styles.css code to fit contact.html
As mentioned in my last commit, here are my changes to the original
styles.css code. I took the background-color declaration for the header
and moved it to the <body> element for full page coverage. I reduced the
margin on the .header-body declaration block. I reduced the font-size of
the <h1> text element. I adjusted the max and min width of the image
element to accommodate a new image with a different aspect ratio. I
removed the maximum character length of the <p> email text element and
added white-space: nowrap to accommodate the new text; simply an email
address. I changed the background-color of the button element to a light
blue shade.