Fix: Fixed relative link to thumbnail
I realised that the html code that I copied over had a relative link in
it, for the IG thumbnail, that I overlooked when I parsed the code for
changes earlier. One full stop fixed everything!

I also removed the comments that I made earlier to keep track of the
changes I was making to styles-pages.css.