Change: Changes to about.html
As I mentioned in my last commit, here are my alterations to the
original contact.html code within about.html. I changed the title
element in the <head> to match the new page, and removed the button
element. I inserted temporary lorumipsum text into the relevant <p>
element and then realised that I was going to need to alter the
.artist-bio selector name, with a corresponding change in css to avoid
contradictory rules with the same specificity, due to using a shared css
styles doc. I did this rather than add a class selector because
id > class.