Feat: Add css declaration block
Added a css declaration block to styles-pages.css targeting the new
artist-about id selector. Because I removed the artist-bio id selector
in about.html and replaced it with a new one, all I had to do to the
styles-pages.css page was add a single css declaration block targeting
minimum and maximum character length of the text element.