Feat: Add hyperlinks to webpages
I added hyperlinks to the <p> element targeted by the .header-title
selector, within about.html and contact.html. This enables navigation
back to the homepage using the header title element as a link. I also
moved the #header-title selector to the <a> link element, since that's
where the text is now nested. I also added the .hyperlink class
selector to the same element to remove the link underline and color.
This worked exactly as I had hoped; the specificity of the id selector
ensured that it had all of the formatting that I wanted, and the class
selector removed the text decoration while similtaneously having the
text color declaration that I didn't want overriden by the specificity
of the id selector.