Feat: Add css declaration block to styles.css
I added a css declaration block targeting the .gallery-image selector.
I defined the height as being 250px, with a corresponding min-height
declaration.

I defined the height and set the width to auto because I
will soon be turning the parent containers for the image elements into
flexboxes with a column flex direction. I set the height parameters as
opposed to width in anticipation of the main axis shifting to vertical.

I also gave the image elements a solid border.