Feat: Add html elements with class selectors
I created a parent container with the .body-container-two class selector
attached. I created two child elements within this container; a text
<div> container with two <p> elements contained within, each with
class selectors; and an <img> element with class selector attached.

This gives me the html framework that I need to structure the next
part of the page body.

I also added the text for the <p> elements and the src link for the
image, effectively dealing with the content for this section as well.