Change: Change padding to margin with new value
I changed the padding-bottom declaration targeting .body-container-one
to margin-bottom, and adjusted the value to 74px.

I did this to center the image elements on the vertical axis. I arrived
at 74px after calculating the total pixel distance from the top of the
image elements to the bottom of the header, then subtracted the pixel
height of the text elements beneath the image elements.