Change: Adjust height and min-height of img class
I slightly adjusted the height and minimum height of the css declaration
block targeting the .gallery-image class selector.

By making the image elements a bit smaller, the page flexes a bit better
in that I would say the page can "survive" being zoomed to 175% in
browser; it can maintain its structural integrity at 175% but elements
start to break out at greater than or equal to 200% zoom. In saying that
the page is somewhat okay at 200% zoom.