Feat: Add <div> <p> <button> <a> <strong> links
I added <div> <p> <button> <a> <strong> elements to index.html. I also
added text to the relevant <p> and <a> links. The <a> link also includes
a href link to Spotify, which is absolute.

This lays the groundwork for adding css to the final section of the body
before crafting the footer.