Change: Remove id's & replace with classes
Once I arrived at placing another button towards the bottom of the home
page, I realised that the way I used selectors to target the button at
the top of the page was inefficient. I have also learned since starting
this project that id selectors should really be reserved for javascript.

Furthermore, I also realised it was more efficient to use the descendant
combinator to target the text within the link element, as seen here
with my changes to the css. I wanted to style the buttons the same,
so I just removed the id selectors and replaced them with a single
button class and a css declaration block targeting the <a> type, nested
in the <button> class.