Feat: Add two css declaration blocks to styles.css
I added two css declaration blocks that both use descendant combinators
to target the .button nested in .body-container-three and the text
embedded in the <a href> element. I gave the margin some padding-left,
and I had to target it this way as not to affect the other button
element sharing the same class selector.

I set the text white-space to nowrap to maintain the integrity of the
button shape when flexing/zooming.