Change: Change id to class selector
I realised that I needed to break up the text in the body of the about
page. Wanting to use multiple <p> elements and taking what I recently
leared about not using ids for css, I changed the #artist-about selector
to a class selector in about.html contact.html and styles-pages.css.