Change: Replace id's with class selectors
As I begin to wrap up this phase of the project and prepare to start
learning Javascript, I considered the possibility of revisiting this
project once I'm ready to apply Javascript to it.

I was recently made aware that id selectors should be used sparingly for
css, and reserved for Javascript. I decided to replace the id selectors
with class selectors by the same name.

I was tempted to remove some selectors entirely and use more nested
targeting to reduce bloat, but rewriting the css at this point feels
redundant. I will however be mindful of this practise going forward.

I also wanted to see if replacing the id's with classes would lead to
specificity issues. The only apparent issue is styling on the
element with the .header-title selector, formerly an id selector, on the
about and contact pages. I will fix this issue in styles-pages.css
before making my next commit.