Fix: Remove text decoration in styles-pages.css
As mentioned previously, I had one fix to make after converting the id
selectors to class selectors; I simply had to remove the underline from
the <a href> element with the .header-title selector.

Initially I thought this to be a case where I would give the .hyperlink
class selector higher specificity by using the descendant combinator,
however I realised that the .hyperlink class had a color declaration
that would override my preferred color declaration in the .header-title
declaration block.

In hindsight I could have written this more efficiently to being with,
but this is a learning exercise first and foremost, so I chalk this up
to a win!

It being written the way it is, I simply added text-decoration: none
to the .header-title declaration block.