Feat: Add head metadata to all html pages
I sent my website on Git Pages to a friend of mine over social media and
I noticed that there was a blank space, clearly intended for a thumbnail
when using instant messaging services.

After doing a bit of research I found html code, intended to be nested
within the <head> element, that should supply a thumbnail along with
other relevant metadata.

Much to my dismay I must push the changes to Github to test this
feature with Git Pages, at least as far as I know. I'm about to make a
push to origin main, so I'm hoping this works straight out the gate!