Change: Edit relative link in head metadata
I sent a link to my Git Pages website to myself on Facebook Messenger
in order to see if the new head metadata is working. It doesn't appear
to be working, and I have several fixes in mind.

In the meantime, I did notice that I forgot to change the relative link
path for about.html and contact.html, seeing as they themselves are in a
sub-directory. I changed each relative link so that they are all
starting at the root using ~/images/etc.

Changing the link in index.html seemed a little pointless, but since I
need to find a fix for the metadata issue I thought it couldn't hurt.
Furthermore, each relative link is the same across all three html files.

After pushing to origin main I will re-test this. It could be a simple
matter of needing to clear my browser cache, to rule out Facebook
reading/storing old metadata, as I did sent the same link to somebody
a week ago. It could even have something to do with the way Github hosts
it's pages.

Otherwise I may have to remove the new head metadata entirely, as a
quick check with ChatGPT seems to confirm that there is nothing wrong
with the code.