Fix: Add new img, add img link to all html pages
In a further attempt to make the Open Graph metadata behave corrently.
I suspect that, due to being metadata, that it may take quite a while to
update on Github Pages or possibly even on Facebook. All I can do for
now is to wait.

In saying that, I have made changes that could potentially be a fix.
I copied one of my existing images and processed it to greatly reduce
file size and pixel resolution, then added it to ~/images. I also
cropped it slightly so that it conforms to the recommended 1.91:1 aspect
ratio, and recommended resolution of 1200px by 630px. I then added the
new relative links to all html pages.

I am hoping that this thumbnail will now display within Facebook
Messenger, but if it doesn't I will leave it alone and test it
again periodically as I suspect this could just be a lag in metadata
properties updating remotely.

I also corrected a mistake in contact.html in which I accidentally
didn't include the /pages directory in the file path, in the og:url
meta element.