Change: Remove gap/add padding, add flex-wrap
I decided to improve the behaviour of the .image-container when zooming.
I removed the gap: 48px from the .image-container css declaration block,
changed the justification of content to space-between, and then added
flex-wrap: wrap so that the images break to a new line when exceeding
the width of the parent container.

To compensate for the removal of the gap, I added 24px of padding to the
left/right of the .image-sub-container elements.

The image elements are behaving much better when zooming in now that
they can break to a new line.