Fix: Add width, flex: auto, & align-self to button
In further efforts to make each element visible/display correctly
at the 200% zoom level, I added a width of 562px to the
.body-container-three class, and yes I did count pixels on this one!
I counted the child element pixel width, including padding, and set the
width according to that.

I also added flex: auto as the objective here was to establish a
flex-basis for better flexing. I also added align-self: center to align
the button on the horizontal cross axis.