Change: Adjust width min-width & padding for image
When I accessed my page on Chrome for Windows, as opposed to Chrome on
Linux where I am developing, I noticed that the image elements in the
gallery were breaking to a new line by default.

Unfortunately, as I am using a VM for Linux, I have to push any changes
to origin main before I can see the results on my Github Page in
Windows Chrome. So fingers crossed that resizing the image elements and
reducing the padding between the child elements will display the way I
want it to on Chrome for Windows on the first try!

Side note: the display is otherwise looking marvellous on Windows Chrome
which is a nice feeling!

Another side note: This is all too similar to producing a music master
in such a way that it sounds good on everything; from mono phone
speakers to high end audiophile systems. All good fun!!