Feat: Add css to styles-pages.css
I added a css declaration block targeting the <html> type selector, and
added justify-content: space-between to the <body> type selector.

The html element in the styles-pages.css is now a flexbox with a column
direction, and justifies content with space between along the vertical
main axis.

This has definitely helped in pushing the footer towards the bottom of
the page for about.html and contact.html.

This now fixed issue isn't apparent with the homepage, presumably owing
to the height of the page which is made taller with all the additional
body content.

I have tested it on several browsers; Windows Chrome, Mozilla, Edge, and
even Safari for Mac OS.

On that note, it is displaying well across the board now; it's time to
move on to learning Javascript!