Change: Add class & css dec. block to about page
In my last commit I added a new image file to ~/images and then replaced
the image source link in about.html.

In this commit I have replaced the .image-header class in about.html
with a new class selector; "about-image". I then added a new css
declaration block to styles-pages.css targeting this new class selector.

This enables styles-sheet.css to target the image elements in both the
about.html and contact.html pages, without any specificity interference
or the need to go back and alter the original class, which is still
targeting contact.html.

The .about-image class is almost identical to the .image-header class,
with the differences being reduced width/flex-basis and reduced
min-width.

These changes were necessary to accommodate the new image file, which
displays in a vertical 9:16 aspect ratio rather than the previous wide
16.9 aspect ratio.

The aspect ratio is the very reason for this change; I produced this
image last night while capturing and editing a short music video.
It occured to me that an image with a 9:16 aspect ratio would display
better, with regards to the text element and it's parent container.